Created SILC Runtime Toolkit git repository Part I.
The new git repository contains the entire history of the SILC
utility library, now named the SILC Runtime Toolkit.  In addition
of that it contains the entire history of the SILC development
since the utility library has been part of that from the very
beginning.  None of the history has been removed.  Only things
that have been removed are unneeded branches and tags.

In this changeset we remove all unnecessary files and directories.