Created SILC Runtime Toolkit git repository Part II.
Added lib/silcutil/silcruntime.h.in, the file that is the main include
file for application programmers.

Added includes/silcompile.h which is included only during compilation
of the Runtime Toolkit.

Changed all files to use the silcruntime.h instead of silc.h header
file.

Made the many other changes in makefiles, configuration scripts and
other files to try to complete the creation of the Runtime git
repository.