Added SILC Buffer Stream API
Buffer stream interface to send and receive buffers.  The benefit of this
interface is that the receiver need not parse buffers from the received
data but each buffer sent is delivered separately to the receiver
callback, even if multiple buffers were received at the same time.  The
length of the buffer is delivered with the data.  The buffer data follows
a 32-bit length field in the stream.